removed QML and QChem section
