+++ demonstrations/guide_to_pennylane_knowing_qiskit.metadata.json
@@ -1,5 +1,5 @@
 {
-    "title": "Your Ultimate Guide to PennyLane if You Know Qiskit",
+    "title": "Your Guide to PennyLane if You Know Qiskit",
     "authors": [
         {
             "id": "isaac_de_vlugt"
